Fix removeData() to accept undefined keys
Incorporates the fix from https://github.com/julianshapiro/velocity/issues/629 which allows removeData to be called without a list of keys to remove all data on the node, as the real jQuery version does. This allows applications to clean up as instructed in the FAQ without having to manually iterate through all the data items that have been set.